backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
python-multipart==0.0.12
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
